+++ apps/hubble-stats/components/KeyMetricItem/index.ts
@@ -1,0 +1,1 @@
+export { default as KeyMetricItem } from './KeyMetricItem';
