+++ apps/hubble-stats/components/KeyMetricItem/types.ts
@@ -1,0 +1,6 @@
+export interface MetricItemProps {
+  title: string;
+  value?: number;
+  changeRate?: number;
+  prefix?: string;
+}
