--- apps/hubble-stats/components/KeyMetricsTable/index.ts
@@ -1,1 +1,0 @@
-export { default as KeyMetricsTable } from './KeyMetricsTable';
