+++ apps/hubble-stats/components/ShieldedAssetsTable/index.ts
@@ -1,0 +1,1 @@
+export { default as ShieldedAssetsTable } from './ShieldedAssetsTable';
