+++ apps/hubble-stats/components/ShieldedAssetsTable/types.ts
@@ -16,19 +16,3 @@ export interface ShieldedAssetsTableProps {
   globalSearchText: string;
   pageSize: number;
 }
-
-export interface ShieldedPoolType {
-  poolSymbol: string;
-  poolAddress: string;
-  poolType: PoolType;
-  token: number;
-  deposits24h: number;
-  tvl: number;
-  chains: string[];
-}
-
-export interface ShieldedPoolsTableProps {
-  data: ShieldedPoolType[];
-  globalSearchText: string;
-  pageSize: number;
-}
