+++ apps/hubble-stats/components/ShieldedPoolsTable/index.ts
@@ -1,0 +1,1 @@
+export { default as ShieldedPoolsTable } from './ShieldedPoolsTable';
