--- apps/hubble-stats/components/ShieldedTables/index.ts
@@ -1,1 +1,0 @@
-export * from './ShieldedTables';
