+++ apps/hubble-stats/containers/KeyMetricsTableContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as KeyMetricsTableContainer } from './KeyMetricsTableContainer';
