+++ apps/hubble-stats/containers/ShieldedTablesContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as ShieldedTablesContainer } from './ShieldedTablesContainer';
