+++ apps/hubble-stats/containers/index.ts
@@ -1,0 +1,2 @@
+export * from './KeyMetricsTableContainer';
+export * from './ShieldedTablesContainer';
